+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v2.2.0
 ------
 * Added a feature to stitch a local stack trace when reading throwables.
 * Added a feature to pass an uncaught exception directly to the handler.
+* Added a feature to transfer bytes from a pipe to an OutputStream.
 
 v2.1.0 (2023-01-06)
 ------